% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenExpBr.R
\name{datagenExpBr}
\alias{datagenExpBr}
\title{Data generation: exponential and normal distributed event-times}
\usage{
datagenExpBr(genpar)
}
\arguments{
\item{genpar}{A vector containing five elements, namely:
\enumerate{
      \item the sample size n,
      \item the backround rate br,
      \item the ADR rate adr,
      \item the relative standard deviation of the ADR generating process rel.sd,
      \item the censor.
       See details for more information.
       }}
}
\value{
A simulated time-event data frame of size n.
}
\description{
Simulation of data (same as for non-Bayesian approach)
}
\details{
Input is a vector with 5 elements
\enumerate{
\item sample size n
\item backround rate br (= observed in the population)
\item Adverse drug reaction rate as prop. of the backround rate
\item rel. sd of the ADR generating process (relative means, the absolut sd = rel.sd*censor)
\item censor (= time point at which the cohort observation period ends)
}

After prespecification of the input, the data simulation works as follows:

The absolut number of cases due to backround causes (other than the adr) is
generated by the binomial distribution with \code{prob = br}.
The absolut number of cases due to the medication is generated by the
binomial distribution with \code{prob = br*adr}. All in all, the expected
number of events/cases within the dataset is
\eqn{n \cdot br (1 + adr)}.

The event-times can have values between zero and censor. Time point zero is the date of
the medication (or start of regular medication) and the censor is the
official end time of the cohort study.

For the br cases, the event-times are generated by an exponential distribution on
the interval [0, censor]. For the adr cases, event-times follow a normal
distribution. The mean of the normal distribution is chosen randomly within
the interval [0,censor]. The standard deviation is defined as
\eqn{ rel.sd \cdot censor}. All generated event-times
\eqn{\leq 0} are set to 1.
}
