% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenpgw.R
\name{datagenpgw}
\alias{datagenpgw}
\title{Data generation: time-event data from pgw distribution with censoring}
\usage{
datagenpgw(genpar)
}
\arguments{
\item{genpar}{A vector containing five elements, namely:
\enumerate{
      \item the sample size n,
      \item the censoring time ct
      \item the scale parameter of the pgw distribution,
      \item the shape parameter of the pgw distribution,
      \item the powershape parameter of the pgw distribution
       }}
}
\value{
A simulated time-event data frame of size n.
}
\description{
Simulates time-event data with cencoring from a specified pgW distribution.
The data simulation works as follows:
The n time observations are generated by the \code{rpgw} function. All time observations
that lie behind the specified censoring time \code{ct} are substituted by the censoring
time itself. The censoring status is defined in a binary vector of length n. To each
observation with time point \eqn{t \leq} \code{ct}, status 1 (event observed) is assigned.
To each observation with time point \eqn{t \geq} \code{ct}, status 0 (no event observed)
is assigned.
}
