% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.fits.R
\docType{data}
\name{first.fits}
\alias{first.fits}
\alias{fst.fit.fgg}
\alias{fst.fit.ggg}
\alias{fst.fit.fll}
\alias{fst.fit.lll}
\title{first fits}
\format{
An object of class \code{stanfit} of dimension 500 x 4 x 3.

An object of class \code{stanfit} of dimension 500 x 4 x 4.

An object of class \code{stanfit} of dimension 500 x 4 x 3.

An object of class \code{stanfit} of dimension 500 x 4 x 4.
}
\usage{
fst.fit.fgg

fst.fit.ggg

fst.fit.fll

fst.fit.lll
}
\description{
Stanfit objects containing the first fit of this kind of function
It is used in future stan model fitting procedures to retrieve the
model structure instead of recompiling it from the C++ code again.
}
\details{
The stanfit objects can be called with

\code{fst.fit.fgg, fst.fit.ggg, fst.fit.fll, fst.fit.lll}.
}
\keyword{datasets}
