% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fgg_function.R
\name{fit.fll}
\alias{fit.fll}
\title{fit prespecified pgw stan models}
\usage{
fit.fll(datstan)
}
\description{
The functions essentially just apply the \code{rstan::stan} command with an already prespecified .stan file.
Within a simulation study, the function can be run in parallel on various prior meta-parameter specifications or datasets.
}
\details{
WICHTIG: warmup (1000) und iter (15000) muss noch angepasst werden!
}
