% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpgw_function.R
\name{plotpgw}
\alias{plotpgw}
\title{Visualization of the power generalized Weibull distribution}
\usage{
plotpgw(scale = 1, shape = 1, powershape = 1, ...)
}
\arguments{
\item{scale}{Scale parameter of the distribution (\eqn{\theta} in formula)}

\item{shape}{Shape parameter (\eqn{\nu} in formula)}

\item{powershape}{Powershape parameter (\eqn{\gamma} in formula)}

\item{...}{optional plot arguments (expept \code{type, lwd, main})}
}
\value{
A plot of the PDF, CDF, hazard and survival function
}
\description{
Visualization of the power generalized Weibull distribution
}
