% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpgw_function.R
\name{qpgw}
\alias{qpgw}
\title{Power generalized Weibull distribution: quantile function}
\usage{
qpgw(p, scale = 1, shape = 1, powershape = 1)
}
\arguments{
\item{p}{percentile in (0,1)}

\item{scale}{Scale parameter of the distribution (\eqn{\theta} in formula)}

\item{shape}{Shape parameter (\eqn{\nu} in formula)}

\item{powershape}{Powershape parameter (\eqn{\gamma} in formula)}
}
\value{
corresponding quantile value
}
\description{
The quantile function is the inverse function of the cumulative distribution
function (here implemented as \code{ppgw}).
}
\details{
The formula is:
\deqn{
q(u) = F^{-1}(u) = \theta \left\{ \left[ 1 - \text{log}(1-u) \right]^{\gamma} - 1 \right\}^{\frac{1}{\nu}}  ]
}
}
