% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpgw_function.R
\name{rpgw}
\alias{rpgw}
\title{Power generalized Weibull distribution: sample generator}
\usage{
rpgw(n, scale = 1, shape = 1, powershape = 1)
}
\arguments{
\item{n}{number of observations}

\item{scale}{Scale parameter of the distribution (\eqn{\theta} in formula)}

\item{shape}{Shape parameter (\eqn{\nu} in formula)}

\item{powershape}{Powershape parameter (\eqn{\gamma} in formula)}
}
\value{
A vector of observations generated from the power generalized Weibull distribution.


The realizations are generated with the analytical inversion method (see for example
\url{https://datasciencegenie.com/how-to-generate-any-random-variable/}).
}
\description{
Power generalized Weibull distribution: sample generator
}
