% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.1.scenario.1.time.R
\name{sim.1.scenario.1.time}
\alias{sim.1.scenario.1.time}
\title{Simulation study: conduct one replication of one scenario}
\usage{
sim.1.scenario.1.time(
  counter,
  scenario.pars,
  save = T,
  path = if (save == T) {
     getwd()
 }
)
}
\arguments{
\item{counter}{what repetition number is it?}

\item{scenario.pars}{Vector with parameter values, that specify the data
generating scenario. See function \code{datagenUnifBr} for
details.}

\item{save}{logical: states whether output is to be saved in a specified file
(T) or in the global environment (F)}

\item{path}{file path specifying where to save the output if save = T}
}
\value{
a data frame object (named \code{raw.sim.table}) containing all fitting and
        posterior sample statistics of the simulation runs (statistics are
        gathered using the \code{stanfit.to.fitstats} and \code{stanfit.to.poststats}
        function).
}
\description{
conducts the data generation, fitting and results output for
one scenario parameter combination one time.
}
\seealso{
for the generating scenario: see function \code{datagenUnifBr} for
                     details.
}
