% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fit.to.1.sample.R
\name{sim.fit.to.1.sample}
\alias{sim.fit.to.1.sample}
\title{Simstudy: apply all models to one specific sample}
\usage{
sim.fit.to.1.sample(survdat)
}
\arguments{
\item{survdat}{time-event-sample}
}
\value{
a data frame with 16 rows. Each row contains statistics for one of the 4x4 prior and model alternatives.
}
\description{
Conducts all steps from data input to simulation output generation including:
}
\details{
- data preparation for stanmodel for all prior assumptions

- fitting all model alternatives

- extracting and returning relevant statistics from stan output:
  - general information "info" (noch meta), eg. prior mean and sds, fitting specifications etc.
  - information about the posterior distributions of nu and gamma "post"
  - credibility intervals for the tests (will be conducted afterwards)
  - 0 to 100 percentiles of the posterior distribution (to roughly estimate the
    probability mass within the region of practical equivalence for the null value)
}
