% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.rep.1.scenario.R
\name{sim.repeat.1.scenario}
\alias{sim.repeat.1.scenario}
\title{Simulation study: replicate one scenario}
\usage{
sim.repeat.1.scenario(counter.vect = 1:100, scenario.pars, path = getwd())
}
\arguments{
\item{scenario.pars}{Vector with parametervalues, that specify the data
generating scenario. See function \code{datagenUnifBr} for
details.}

\item{path}{file path specifying where to save the output if save = T}

\item{reps}{number of repeated simulations for one scenario}

\item{save}{logical: states whether output is to be saved in a specified file
(T) or in the global environment (F)}
}
\value{
a data frame object (named \code{raw.sim.table}) containing all fitting and
        posterior sample statistics of the simulation runs (statistics are
        gathered using the \code{stanfit.to.fitstats} and \code{stanfit.to.poststats}
        function).
                     generating scenario. See function \code{datagenUnifBr} for
                     details.
}
\description{
repeat the data generation and fitting and results output for
one scenario parameter combination
}
