% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.to.fitstats.R
\name{stanfit.to.fitstats}
\alias{stanfit.to.fitstats}
\title{extract meta and prior statistics from stanfit object}
\usage{
stanfit.to.fitstats(stanfit.object, stan.dat, assumed.pr, adr.gen.mean.time)
}
\arguments{
\item{stanfit.object}{the estimated stan model output}

\item{stan.dat}{the data inserted for model fitting}
}
\value{
Information about the model fitting, such as:
\item{stancode.name}{to identify the used model}
\item{adr.assumption}{to identify which adr case is assumed a priori}
\item{adr.gen.mean.time}{to reflect, which adr prior assumption
      would be most fitting theoretically}
\item{th.pr.mean}{prior mean of parameter theta}
\item{th.pr.sd}{prior standard deviation of parameter theta}
\item{nu.pr.mean}{prior mean of parameter nu}
\item{nu.pr.sd}{prior standard deviation of parameter nu}
\item{ga.pr.mean}{prior mean of parameter gamma}
\item{ga.pr.sd}{prior standard deviation of parameter gamma}
\item{run.min}{running time in minutes}
}
\description{
The function extracts information about the model fitting and prior specifications for the parameters to be estimated.
}
