% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.to.metastats.R
\name{stanfit.to.metastats}
\alias{stanfit.to.metastats}
\title{extract meta  and prior statistics from stanfit object}
\usage{
stanfit.to.metastats(stanfit.object, stan.dat)
}
\arguments{
\item{stanfit.object}{the estimated stan model output}

\item{stan.dat}{the data inserted for model fitting}
}
\value{
Information about the model fitting, such as:
- model.seed: the seed was generated and set to make the posterior samples reproducible
- stancode.name: to identify the used model
- th.pr.mean: prior mean of parameter theta
}
\description{
The function extracts iformation about the model fitting and prior specifications for the parameters to be estimated.
}
