% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.to.poststats.R
\name{stanfit.to.poststats}
\alias{stanfit.to.poststats}
\title{extract posterior summary statistics and credibility intervals}
\usage{
stanfit.to.poststats(stanfit.object, cred.niveaus = seq(0.5, 0.95, by = 0.05))
}
\arguments{
\item{stanfit.object}{the estimated stan model output}

\item{cred.niveaus}{}
}
\value{
Information about the posterior distributions stored in a vector
        consisting of the following entry parts first for parameter nu followed by the same statistics regarding
        parameter gamma:

\item{nu/ga.post.stats}{Summary statistics from the stanfit object about the posterior distribution, namely \code{mean, se_mean, sd},
                 and bayesian convergence diagnostic measures, namely \cod{n_eff, Rhat}}
\item{nu/ga.eti}{equitailed posterior (credibility) intervals specified by their lower and upper boundaries;
                 one interval is derived for each specified credibility niveau}
\item{nu/ga.hdi}{highest posterior density intervals specified by their lower and upper boundaries;
                 one interval is derived for each specified credibility niveau}
\item{nu/ga.per}{percentiles for the posterior density in order to estimate the incorporated probability mass within a
                 specified region of practical equivalence (rope) for the null value of the statistical test that will be
                 conducted after the fitting procedure}

 @details The storing in one long vector for each parameter is motivated by the
 goal to conduct a simulation study, where all relevant statistics per run ars stored columnwise
 and the repeated runs for one data scenario are store rowwise.
}
\description{
Given a stanfit object, the function extracts statistics about the posterior
distribution of the shape parameters nu and gamma.
These shall give an overview and provide the basis for the adr signal detection
testing in one vector.
}
