% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.to.priorstats.R
\name{stanfit.to.priorstats}
\alias{stanfit.to.priorstats}
\title{extract meta  and prior statistics from stanfit object}
\usage{
stanfit.to.priorstats(stanfit.object, stan.dat)
}
\arguments{
\item{stanfit.object}{the estimated stan model output}

\item{stan.dat}{the data inserted for model fitting}
}
\value{
Information about the model fitting, such as:
\item{model.seed}{the seed was generated and set to make the posterior samples reproducible}
\item{stancode.name}{to identify the used model}
\item{th.pr.mean}{prior mean of parameter theta}
\item{th.pr.sd}{prior standard deviation of parameter theta}
\item{nu.pr.mean}{prior mean of parameter nu}
\item{nu.pr.sd}{prior standard deviation of parameter nu}
\item{ga.pr.mean}{prior mean of parameter gamma}
\item{ga.pr.sd}{prior standard deviation of parameter gamma}
\item{run.min}{running time in minutes}
}
\description{
The function extracts iformation about the model fitting and prior specifications for the parameters to be estimated.
}
