% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodel.pgw.R
\docType{stanmodel}
\name{stanmodel.pgw}
\alias{stanmodel.pgw}
\alias{stanmodel.pgw.fix.gam.gam.repar}
\alias{stanmodel.pgw.gam.gam.gam.repar}
\alias{stanmodel.pgw.fix.logn.logn.repar}
\alias{stanmodel.pgw.logn.logn.logn.repar}
\title{stan Models pgw distribution}
\format{
An object of class \code{stanmodel} of length 1.

An object of class \code{stanmodel} of length 1.

An object of class \code{stanmodel} of length 1.

An object of class \code{stanmodel} of length 1.
}
\usage{
rstan::sampling(object = stanmodel.pgw.fix.gam.gam.repar, data = ..., ...)
rstan::sampling(object = stanmodel.pgw.gam.gam.gam.repar, data = ..., ...)
rstan::sampling(object = stanmodel.pgw.fix.logn.logn.repar, data = ..., ...)
rstan::sampling(object = stanmodel.pgw.logn.logn.logn.repar, data = ..., ...)

stanmodel.pgw.fix.gam.gam.repar

stanmodel.pgw.gam.gam.gam.repar

stanmodel.pgw.fix.logn.logn.repar

stanmodel.pgw.logn.logn.logn.repar
}
\description{
Prepared stanmodels allow Bayesian fitting of a
power generalized Weibull model with log-normal prior assumption for each parameter.
}
\details{
The data argument in the \code{rstan::sampling} command needs to be a time-event
dataset reformatted with the \code{bADRfcts::survdat2pgwstanmodeldat} function.

For information on further arguments in the fitting function see \code{?rstan::sampling}.
}
\keyword{datasets}
\keyword{pgw}
\keyword{stanmodel}
