% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdat.R
\docType{data}
\name{testdat}
\alias{testdat}
\title{test data for development of simstudy code}
\format{
An object of class \code{data.frame} with 1000 rows and 2 columns.
}
\usage{
testdat
}
\description{
generated with \code{datagenpgw} command with
scale =3, shape = 5, powershape = 15.
}
\keyword{datasets}
